Put flask handler into main module (will have to break out if making whole site.
