+++ runserver.py
@@ -1,3 +1,3 @@
-from flask_rest_service import app
+from dmaid import app
 if __name__ == '__main__':
     app.run(debug=True)
